remove eadid memoization
The sequencing of EADID extraction is fiddly, because one must
capture the EADID from the EAD XML file before actually deleting
the EAD file from the repo.  However, one shouldn't finish writing
the commit message if the delete was unsuccessful.

In the last version of the code, I memoized EADID, but this was
a bad idea because there can be multiple delete operations in
one processing pass, and all deleted EADs but the first would have
incorrect commit messages.

Therefore, I made EADID extraction a separate method that captures
the EADID and stored is in an instance variable.

The extract_eadid method is called before the EAD XML file is
removed from the repo.